agents/dispensary_scraper/requirements.txt
# Core Dependencies
pydantic-ai>=0.0.12
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web Scraping Dependencies
playwright>=1.40.0

# Database Dependencies  
snowflake-connector-python[pandas]>=3.6.0
pandas>=2.1.0

# Environment Management
python-dotenv>=1.0.0

# Async Support
asyncio>=3.4.3
nest-asyncio>=1.5.8

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Dependencies
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Logging and Monitoring
structlog>=23.2.0

# HTTP Client (for potential API integrations)
httpx>=0.25.0

# Type Hints
types-requests>=2.31.0